refactor(github): extract GitHub utils into provider + rename factory
- Add repo-providers/github-provider.mjs with createGithubProvider factory
- Keep backward compat via github-utils.mjs re-export and alias createGithubUtils
- Update workflow and CLI to use repoProvider variable naming
- Move and rename tests to repo-providers/tests/github-provider.test.mjs
- Add validation test for missing githubToken
- Adjust jest config for new test location and coverage exclusions
- Remove legacy duplicated test content (stub removed)
- Update integration tests to assert new factory and alias
