+++ src/gen-pr.mjs
@@ -19,7 +19,7 @@ import {
 } from "./ai/chatgpt.mjs";
 import { executePRWorkflow } from "./workflow.mjs";
 import { getCurrentBranch } from "./git-utils.mjs";
-import { createGithubUtils } from "./github-utils.mjs";
+import { createGithubProvider } from "./repo-providers/github-provider.mjs";
 
 const argv = minimist(process.argv.slice(2), {
     alias: {
@@ -263,7 +263,8 @@ const main = async () => {
 
     // Call function from workflow
     // Create GitHub utils bound to token and pass to workflow
-    const githubUtils = createGithubUtils({ githubToken: config.githubToken });
+    // Instantiate repository provider (GitHub implementation)
+    const repoProvider = createGithubProvider({ githubToken: config.githubToken });
 
     await executePRWorkflow(
         sourceBranch,
@@ -274,7 +275,7 @@ const main = async () => {
         remoteSourceBranch,
         remoteTargetBranch,
         upstreamRemoteName || remoteName,
-        githubUtils
+        repoProvider
     );
 };
